qtypetraits.h: add is_signed/is_unsigned/not_/not_c predicates
is_signed/is_unsigned check whether the type argument is signed
or unsigned, resp., and will be used in QFlags to select the
correct underlying integer type.

not_ is used in the implementation of is_signed

not_c is for completeness (version of not_ that takes a bool
instead of an integral_constant).

Change-Id: I77cc445e8c6cf3181336505c9c13478fba3e7890
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>